完善单聊 完善token临时方案,后面会换成JWT 前端页面修改 部分代码优化 模板自动加载

+++ models/chat_contact.go
@@ -2,8 +2,8 @@ package models
 
 type ChatContact struct {
 	Id        int64  `xorm:"pk autoincr comment('主键') BIGINT(20)"`
-	UserId    int    `xorm:"not null comment('用户ID,记录所属人ID') INT(11)"`
-	TargetId  int    `xorm:"not null comment('对方用户ID') INT(11)"`
+	UserId    int64  `xorm:"not null comment('用户ID,记录所属人ID') INT(11)"`
+	TargetId  int64  `xorm:"not null comment('对方用户ID') INT(11)"`
 	Type      int    `xorm:"not null comment('类型(用户对用户,用户对群)') TINYINT(4)"`
 	Remark    string `xorm:"not null default '' comment('备注') VARCHAR(255)"`
 	Created   int    `xorm:"not null default 0 comment('创建时间') INT(15)"`
